Refactor: separate components based on pages
