+++ src/components/FolderPage/FolderButtons.css
@@ -1,0 +1,14 @@
+.folderButton {
+  background-color: white;
+  border: 1px solid var(--primary-color);
+  padding: 8px 12px;
+  border-radius: 5px;
+  cursor: pointer;
+}
+
+.folderButton:hover,
+.folderButton:active,
+.folderButton:focus {
+  background-color: var(--primary-color);
+  color: white;
+}
